all SO enemies have been set up, sell/buy menus fully functional, waves and spawners seem to be working correctly, the player can take damage and the skip button work as intender at the moment everything is very rudimentary, there is not wave time, and player HP is only visible in the inspector. will refactor the hexgridmanager as at the moment it handles to many tasks and it seems a bit complex to read.
first level fully functional

+++ Assets/Scripts/ScriptableObjects/Enemies/EnemyInfo.cs
@@ -7,7 +7,7 @@ public class EnemyInfo : ScriptableObject
 
     public int maxHealth;
     public float moveSpeed;
-    public int dangerLevel;
+    public float dangerLevel;
 
 
     public bool isAerial;
